Split e4-debug.h into appropriate include files
Splits the proof-of-concept e4-debug.h header file into multiple public
and private header files and more appropriately resemble the projected
structure. Adds various TODO and FIXME comments as necessary.

Preserves the e4-debug.h header file for the time being, as it includes
stdio.h, which is currently useful. In the future, perhaps it would be a
good idea to define some e4__DEBUG_PRINTF macro to nothing in e4.h and
then define it to printf in e4-debug.h, making it so that by including
or not including e4-debug.h it could be possible to quickly toggle debug
output.