Add return stack coherence tracking
Adds handling to ensure the return stack has not become corrupt while
executing threaded code, throwing the appropriate exceptions when it
has. This seems to work correctly and has even helped to uncover a
couple of bugs, but it still requires additional test coverage, for
which the TODO has been updated accordingly.